Tar files not yet deleted
